pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ccas"
version = "0.1.0"
description = "CCAS: city research, paper extraction, and relevance scoring"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "pandas",
    "numpy",
    "openai",
    "dropbox",
    "requests",
    "scikit-learn",
    "supabase",
    "PyMuPDF",
]

[tool.setuptools.packages.find]
include = ["ccas*"]
